- Temporary Commit 2025.03.04.0100

+++ app/src/main/java/com/aetherized/smartfinance/features/finance/domain/model/DailySummary.kt
@@ -5,6 +5,6 @@ import java.time.LocalDate
 data class DailySummary(
     val date: LocalDate,
     val transactions: List<Transaction>,
-    val income: Int,   // or Double if needed
-    val expense: Int
+    val income: Double,   // or Double if needed
+    val expense: Double
 )
